+++ test/user.clj
@@ -1,5 +1,7 @@
 (ns user
-  (:require [net.lewisship.trace :as trace]))
+  (:require [net.lewisship.trace :as trace]
+    ;; Enable (is (match? ..)):
+            matcher-combinators.clj-test))
 
 (trace/setup-default)